lemm_Counter 02

+++ src/main/java/searchengine/services/LemmaService.java
@@ -4,5 +4,5 @@ import java.io.IOException;
 import java.util.Map;
 
 public interface LemmaService {
-    public Map<String,Integer> LemmaMap(String s) throws IOException;
+    public Map<String,Double> LemmaMap(String s) throws IOException;
 }
